Fix loading of files and add a Swedish text with specs
