--- spec/fullTextTests/testTexts/englishPaper2.js
@@ -1,133 +1,0 @@
-import Paper from "../../../src/values/Paper.js";
-import loadText from "../../../src/helpers/loadFullTextPapers";
-
-const language = "en";
-const name = "englishPaper2";
-
-const paper = new Paper( loadText( language, name ), {
-	keyword: "Google Search Console data",
-	synonyms: "Google Click Through Rate, Google CTR",
-	description: "Google Search Console has a totally new interface! And new features: e.g. 16 months of data. Annelieke explains what you can do with it!",
-	title: "Annelieke's Analytics: 16 months of Google Search Console data",
-	titleWidth: 450,
-	locale: "en_EN",
-	url: "https://yoast.com/16-months-of-google-search-console-data/",
-} );
-
-const expectedResults = {
-	introductionKeyword: {
-		score: 6,
-		resultText: "<a href='https://yoa.st/33e' target='_blank'>Keyphrase in introduction</a>:Your keyphrase or its synonyms appear in the first paragraph of the copy, but not within one sentence. <a href='https://yoa.st/33f' target='_blank'>Fix that</a>!",
-	},
-	keyphraseLength: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/33i' target='_blank'>Keyphrase length</a>: Good job!",
-	},
-	keywordDensity: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/33v' target='_blank'>Keyphrase density</a>: 0.7%. This is great!",
-	},
-	metaDescriptionKeyword: {
-		score: 3,
-		resultText: "<a href='https://yoa.st/33k' target='_blank'>Keyphrase in meta description</a>: The meta description has been specified, but it does not contain the keyphrase. <a href='https://yoa.st/33l' target='_blank'>Fix that</a>!",
-	},
-	metaDescriptionLength: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34d' target='_blank'>Meta description length</a>: Well done!",
-	},
-	subheadingsKeyword: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/33m' target='_blank'>Keyphrase in subheading</a>: Your subheading reflects the topic of your copy. Good job!",
-	},
-	textCompetingLinks: {
-		score: 0,
-		resultText: "",
-	},
-	textImages: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/33c' target='_blank'>Image alt attributes</a>: Some images on this page contain alt attributes with words from your keyphrase! Good job!",
-	},
-	textLength: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34n' target='_blank'>Text length</a>: The text contains 908 words. Good job!",
-	},
-	externalLinks: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34f' target='_blank'>Outbound links</a>: Good job!",
-	},
-	internalLinks: {
-		score: 3,
-		resultText: "<a href='https://yoa.st/33z' target='_blank'>Internal links</a>: No internal links appear in this page, <a href='https://yoa.st/34a' target='_blank'>make sure to add some</a>!",
-	},
-	titleKeyword: {
-		score: 6,
-		resultText: "<a href='https://yoa.st/33g' target='_blank'>Keyphrase in title</a>: The exact match of the keyphrase appears in the SEO title, but not at the beginning. <a href='https://yoa.st/33h' target='_blank'>Try to move it to the beginning</a>.",
-	},
-	titleWidth: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34h' target='_blank'>SEO title width</a>: Good job!",
-	},
-	urlKeyword: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/33o' target='_blank'>Keyphrase in slug</a>: More than half of your keyphrase appears in the slug. That's great!",
-	},
-	urlLength: {
-		score: 6,
-		resultText: "<a href='https://yoa.st/35b' target='_blank'>Slug too long</a>: the slug for this page is a bit long. <a href='https://yoa.st/35c' target='_blank'>Shorten it</a>!",
-	},
-	urlStopWords: {
-		score: 5,
-		resultText: "<a href='https://yoa.st/34p' target='_blank'>Slug stopwords</a>: The slug for this page contains a stop word. <a href='https://yoa.st/34q' target='_blank'>Remove it</a>!",
-	},
-	keyphraseDistribution: {
-		score: 6,
-		resultText: "<a href='https://yoa.st/33q' target='_blank'>Keyphrase distribution</a>: Uneven. Some parts of your text do not contain the keyphrase or its synonyms. <a href='https://yoa.st/33u' target='_blank'>Distribute them more evenly</a>.",
-	},
-	fleschReadingEase: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34r' target='_blank'>Flesch Reading Ease</a>: The copy scores 75.1 in the test, which is considered fairly easy to read. Good job!",
-	},
-	subheadingsTooLong: {
-		score: 3,
-		resultText: "<a href='https://yoa.st/34x' target='_blank'>Subheading distribution</a>: " +
-		"2 sections of your text are longer than 300 words and are not separated by any subheadings." +
-		" <a href='https://yoa.st/34y' target='_blank'>Add subheadings to improve readability</a>.",
-	},
-	textParagraphTooLong: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/35d' target='_blank'>Paragraph length</a>: None of the paragraphs are too long. Great job!",
-	},
-	textSentenceLength: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34v' target='_blank'>Sentence length</a>: Great!",
-	},
-	textTransitionWords: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34z' target='_blank'>Transition words</a>: Well done!",
-	},
-	passiveVoice: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: You're using enough active voice. That's great!",
-	},
-	textPresence: {
-		score: 0,
-		resultText: "",
-	},
-	sentenceBeginnings: {
-		score: 9,
-		resultText: "<a href='https://yoa.st/35f' target='_blank'>Consecutive sentences</a>: There is enough variety in your sentences. That's great!",
-	},
-};
-
-export {
-	name,
-	paper,
-	expectedResults,
-};
-
-export default {
-	name: name,
-	paper: paper,
-	expectedResults: expectedResults,
-};
-
