+++ spec/fullTextTests/testTexts/index.js
@@ -1,11 +1,18 @@
-import fs from "fs";
+// English papers
 
-import englishPaper1 from "./englishPaper1";
-import englishPaper2 from "./englishPaper2";
-import englishPaper3 from "./englishPaper3";
+import englishPaper1 from "./en/englishPaper1";
+import englishPaper2 from "./en/englishPaper2";
+import englishPaper3 from "./en/englishPaper3";
+
+// Swedish papers
+
+import swedishPaper1 from "./sv/swedishPaper1";
+
 
 export default [
 	englishPaper1,
 	englishPaper2,
 	englishPaper3,
+	swedishPaper1,
 ];
+
