+++ spec/fullTextTests/testTexts/sv/swedishPaper1.html
@@ -10,4 +10,5 @@
 </p><p>Familjens förmögenhet ökade och tillät pappa Daniel att köpa flera tomter vid <a href="/wiki/Skurusundet" title="Skurusundet">Skurusundet</a> på <a href="/wiki/V%C3%A4rmd%C3%B6" title="Värmdö">Värmdö</a> utanför Stockholm. Där, i nära anslutning till <a href="/wiki/Eken%C3%A4s" title="Ekenäs">Ekenäs</a> brygga och <a href="/wiki/Waxholm" class="mw-redirect" title="Waxholm">Waxholmsbåtarna</a>, uppfördes familjens <i>Ekefors</i> med tolvrumsvilla och rymlig målarateljé. Daniel Nordström hade fått det hedrande uppdraget att måla 21 landskapsbilder i taket på nyrestaurerade salongen Grand Restaurang National, nuvarande <a href="/wiki/Nalen" title="Nalen">Nalen</a>, på Regeringsgatan 74.<sup id="cite_ref-FOOTNOTEBremmer201752-53_17-1" class="reference"><a href="#cite_note-FOOTNOTEBremmer201752-53-17"><span class="cite-reference-link-bracket">[</span>16<span class="cite-reference-link-bracket">]</span></a></sup>
 </p>
 
-"Ester Belenda Nordström" by Wikipedia is licensed under CC-BY-SA 2.0"
+<!-- "Ester Blenda Nordström" by Wikipedia is licensed under CC-BY-SA 2.0 (https://creativecommons.org/licenses/by-sa/2.0/) -->
+
