+++ src/helpers/loadFullTextPapers.js
@@ -3,12 +3,11 @@ import path from "path";
 
 /**
  *
- * @param language
- * @param textName
- * @returns {*}
+ * @param {string} language The language of the text
+ * @param {string} textName The name of the text
+ * @returns {*} The test text from an html fle
  */
-export default function ( language, textName )
-{
+export default function( language, textName ) {
 	return fs.readFileSync(
 		path.join( "./spec/fullTextTests/testTexts", language, `${textName}.html` ),
 		{ encoding: "UTF8" } );
